refactor(employer): 정산 화면 features 모듈 분할 (Phase 5)
- src/features/employer/settlements/: settlementCalc(계산·파생 헬퍼) + useStaffSettlementsHandlers(핸들러 10종) + TabHeader
- src/components/employer/settlements/SettlementModals: 하단 모달 다발 props 컴포넌트화
- 핸들러 클로저 의존을 훅 인자로 주입, useCallback deps 바이트 동일
- 진입점 settlements.tsx(819→233)는 _layout 의존+훅 호출+가드+조립만, default export 불변
- 정산 계산·토스트 문구·dark: 불변 (tsc 0, test 4286 pass)
